refactor(tokenRefresh): extract 12 per-provider refresh functions to co-located files
Splits open-sse/services/tokenRefresh.ts (1,996 -> 887 LOC, -56%) by
moving 12 per-provider refresh*Token functions (1,109 LOC total) to
co-located files under open-sse/services/tokenRefresh/providers/.

## What moved (12 functions, 1,109 LOC extracted)

| File | LOC | Function |
|---|---|---|
| providers/windsurf.ts   | 106  | refreshWindsurfToken |
| providers/cline.ts      | 69   | refreshClineToken |
| providers/kimiCoding.ts | 113  | refreshKimiCodingToken |
| providers/gitlabDuo.ts  | 92   | refreshGitLabDuoToken |
| providers/claudeOAuth.ts| 59   | refreshClaudeOAuthToken |
| providers/google.ts     | 61   | refreshGoogleToken |
| providers/qwen.ts       | 81   | refreshQwenToken |
| providers/codex.ts      | 86   | refreshCodexToken |
| providers/kiro.ts       | 221  | refreshKiroToken |
| providers/qoder.ts      | 59   | refreshQoderToken |
| providers/github.ts     | 48   | refreshGitHubToken |
| providers/copilot.ts    | 114  | refreshCopilotToken |

## What stayed in tokenRefresh.ts (887 LOC)

All 13 generic helpers used by the per-provider functions:
buildFormParams, extractOAuthErrorCode, getRefreshLeadMs, readRefreshErrorBody,
refreshAccessToken, refreshWithRetry, recordSuccess, recordFailure,
withTimeout, getRefreshCacheKey, runWithOnPersist, getActiveOnPersist,
cleanupRotationMap, lookupRotation, recordRotation.

## Why this is safe

Each per-provider function is a self-contained module-level async function
that takes a TokenRefreshRequest and returns Promise<TokenRefreshResult>.
The 4 functions that need shared helpers (claudeOAuth, qoder, github,
copilot) explicitly import them from '../tokenRefresh' - the import
graph is clean and acyclic.

## Pattern matches PR-#4609 (imageGeneration split) + PR-#4609 (providers split)

Same mechanical refactor pattern:
1. Identify leaf functions with shared-helper deps mapped (allowed up to 2 helpers)
2. Move to co-located file with import block
3. Add re-export from parent
4. Verify references unchanged

## Verification performed locally

- Re-export present in open-sse/services/tokenRefresh.ts: 12 lines
- Per-provider refresh functions still in main file: 0 (moved)
- node scripts/check/check-env-doc-sync.mjs: 'Env / docs contract is in sync'
- node scripts/check/check-fabricated-docs.mjs --strict: 'No fabricated
  API/env/CLI/hook/file references found'
- node scripts/check/check-docs-sync.mjs: 'PASS'
- node scripts/check/check-db-rules.mjs: 'OK'

## Refs

- Builds on PR-#4609 (imageGeneration split) and the providers.ts split
  in this same branch - same leaf-block pattern, same review shape
- Mirrors PR-#4381 (combos split) and PR-#4480 (chatLogHelpers extraction)
- Continues work on issue #4425 (provider-config drift crashes)
