+++ open-sse/services/tokenRefresh.ts
@@ -8,6 +8,21 @@ import { serializeRefresh } from "./refreshSerializer.ts";
 import { WINDSURF_CONFIG } from "@/lib/oauth/constants/oauth";
 import { buildGitLabOAuthEndpoints, resolveGitLabOAuthBaseUrl } from "@/lib/oauth/gitlab";
 
+// --- Per-provider refresh functions (extracted to co-located files in PR-#4609-batch) ---
+export { refreshWindsurfToken } from "./tokenRefresh/providers/windsurf";
+export { refreshClineToken } from "./tokenRefresh/providers/cline";
+export { refreshKimiCodingToken } from "./tokenRefresh/providers/kimiCoding";
+export { refreshGitLabDuoToken } from "./tokenRefresh/providers/gitlabDuo";
+export { refreshClaudeOAuthToken } from "./tokenRefresh/providers/claudeOAuth";
+export { refreshGoogleToken } from "./tokenRefresh/providers/google";
+export { refreshQwenToken } from "./tokenRefresh/providers/qwen";
+export { refreshCodexToken } from "./tokenRefresh/providers/codex";
+export { refreshKiroToken } from "./tokenRefresh/providers/kiro";
+export { refreshQoderToken } from "./tokenRefresh/providers/qoder";
+export { refreshGitHubToken } from "./tokenRefresh/providers/github";
+export { refreshCopilotToken } from "./tokenRefresh/providers/copilot";
+
+
 // Default token expiry buffer (refresh if expires within 5 minutes).
 // Used as fallback for providers without an explicit lead time in
 // REFRESH_LEAD_MS below.
@@ -356,1115 +371,6 @@ export async function refreshAccessToken(
  * expires and refresh is a no-op returning the same token.
  * If the token is a Firebase ID token (device-code flow), it expires after
  * ~1 hour and can be refreshed with the stored Firebase refresh token.
- */
-export async function refreshWindsurfToken(
-  refreshToken: string,
-  providerSpecificData: Record<string, unknown> | null | undefined,
-  log: RefreshLogger,
-  proxyConfig: unknown = null
-) {
-  if (!refreshToken) {
-    log?.warn?.(
-      "TOKEN_REFRESH",
-      "No refresh token stored for Windsurf — token may be a long-lived API key"
-    );
-    return null;
-  }
-
-  const authMethod = (providerSpecificData?.authMethod as string) || "import";
-
-  // Long-lived Codeium API keys (import flow) have no expiry — nothing to refresh.
-  if (authMethod === "import") {
-    log?.debug?.("TOKEN_REFRESH", "Windsurf import token is long-lived — no refresh needed");
-    return null;
-  }
-
-  // Firebase STS refresh for browser-flow tokens.
-  // Resolves via WINDSURF_CONFIG.firebaseApiKey, which honors the
-  // WINDSURF_FIREBASE_API_KEY env override and falls back to the embedded
-  // public default in publicCreds.ts. See docs/security/PUBLIC_CREDS.md.
-  const firebaseApiKey = WINDSURF_CONFIG.firebaseApiKey || "";
-  if (!firebaseApiKey) {
-    log?.warn?.(
-      "TOKEN_REFRESH",
-      "Windsurf Firebase API key unavailable — skipping Firebase token refresh"
-    );
-    return null;
-  }
-  const tokenUrl = `https://securetoken.googleapis.com/v1/token?key=${firebaseApiKey}`;
-
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(tokenUrl, {
-        method: "POST",
-        headers: { "Content-Type": "application/x-www-form-urlencoded" },
-        body: buildFormParams({ grant_type: "refresh_token", refresh_token: refreshToken }),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Windsurf Firebase token", {
-        status: response.status,
-        error: errorText.slice(0, 200),
-      });
-
-      // Firebase STS returns structured errors. Detect unrecoverable token states.
-      try {
-        const fbError = JSON.parse(errorText);
-        const fbCode =
-          typeof fbError?.error?.message === "string"
-            ? fbError.error.message
-            : typeof fbError?.error === "string"
-              ? fbError.error
-              : null;
-        if (
-          typeof fbCode === "string" &&
-          (fbCode.includes("USER_DISABLED") ||
-            fbCode.includes("TOKEN_EXPIRED") ||
-            fbCode.includes("INVALID_REFRESH_TOKEN") ||
-            fbCode.includes("USER_NOT_FOUND"))
-        ) {
-          log?.error?.(
-            "TOKEN_REFRESH",
-            "Windsurf Firebase token is permanently invalid. Re-authentication required.",
-            {
-              fbCode,
-            }
-          );
-          return { error: "unrecoverable_refresh_error", code: fbCode };
-        }
-      } catch {
-        // not JSON — fall through
-      }
-
-      return null;
-    }
-
-    const data = await response.json();
-    const expiresIn = parseInt(data.expires_in ?? "3600", 10);
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Windsurf Firebase token", {
-      expiresIn,
-      hasNewIdToken: !!data.id_token,
-    });
-
-    return {
-      accessToken: data.id_token,
-      refreshToken: data.refresh_token || refreshToken,
-      expiresIn,
-    };
-  } catch (error) {
-    log?.error?.(
-      "TOKEN_REFRESH",
-      `Network error refreshing Windsurf token: ${error instanceof Error ? error.message : String(error)}`
-    );
-    return null;
-  }
-}
-
-export async function refreshClineToken(refreshToken, log, proxyConfig: unknown = null) {
-  const endpoint = PROVIDERS.cline?.refreshUrl;
-  if (!endpoint) {
-    log?.warn?.("TOKEN_REFRESH", "No refresh URL configured for Cline");
-    return null;
-  }
-
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(endpoint, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Accept: "application/json",
-        },
-        body: JSON.stringify({
-          refreshToken,
-          grantType: "refresh_token",
-          clientType: "extension",
-        }),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Cline token", {
-        status: response.status,
-        error: errorText,
-      });
-      const code = extractOAuthErrorCode(errorText);
-      if (code === "invalid_grant" || code === "invalid_request") {
-        return { error: "unrecoverable_refresh_error", code };
-      }
-      return null;
-    }
-
-    const payload = await response.json();
-    const data = payload?.data || payload;
-    const expiresAtIso = data?.expiresAt;
-    const expiresIn = expiresAtIso
-      ? Math.max(1, Math.floor((new Date(expiresAtIso).getTime() - Date.now()) / 1000))
-      : undefined;
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Cline token", {
-      hasNewAccessToken: !!data?.accessToken,
-      hasNewRefreshToken: !!data?.refreshToken,
-      expiresIn,
-    });
-
-    return {
-      accessToken: data?.accessToken,
-      refreshToken: data?.refreshToken || refreshToken,
-      expiresIn,
-    };
-  } catch (error) {
-    log?.error?.("TOKEN_REFRESH", `Network error refreshing Cline token: ${error.message}`);
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for Kimi Coding OAuth tokens.
- * Uses custom X-Msh-* headers required by Kimi OAuth API.
- *
- * Uses a stable device_id from providerSpecificData (stored at login) to avoid
- * anti-bot detection from ephemeral IDs. If absent, derives a deterministic ID
- * from the refresh token hash so it is at least stable across refreshes for the
- * same token.
- */
-export async function refreshKimiCodingToken(
-  refreshToken: string,
-  providerSpecificData: Record<string, unknown> | null | undefined,
-  log: RefreshLogger,
-  proxyConfig: unknown = null
-) {
-  const endpoint = PROVIDERS["kimi-coding"]?.refreshUrl || PROVIDERS["kimi-coding"]?.tokenUrl;
-  if (!endpoint) {
-    log?.warn?.("TOKEN_REFRESH", "No refresh URL configured for Kimi Coding");
-    return null;
-  }
-
-  // Prefer stable device_id persisted at login time; fall back to a
-  // deterministic hash of the refresh token so it is at least consistent
-  // across refreshes for the same session.
-  const stableDeviceId =
-    (providerSpecificData?.deviceId as string) ||
-    pbkdf2Sync(refreshToken, "kimi-device-id", 1000, 16, "sha256").toString("hex");
-
-  const platform = "kimi_cli";
-  const version = process.env.KIMI_CLI_VERSION || "1.36.0";
-
-  // Build device model string matching the format from providers/kimi-coding.ts.
-  // open-sse is a portable workspace — use process.platform/arch (always available in Node).
-  const osTypeStr = typeof process !== "undefined" ? process.platform : "unknown";
-  const archStr = typeof process !== "undefined" ? process.arch : "unknown";
-  const deviceModel = [osTypeStr, archStr].filter(Boolean).join(" ");
-
-  try {
-    const params = new URLSearchParams({
-      grant_type: "refresh_token",
-      refresh_token: refreshToken,
-      client_id: PROVIDERS["kimi-coding"]?.clientId || "",
-    });
-
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(endpoint, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-          Accept: "application/json",
-          "X-Msh-Platform": platform,
-          "X-Msh-Version": version,
-          "X-Msh-Device-Model": (providerSpecificData?.deviceModel as string) || deviceModel,
-          "X-Msh-Device-Id": stableDeviceId,
-          // These headers match getKimiOAuthHeaders() in providers/kimi-coding.ts.
-          // They're derived at runtime from os module calls; use safe fallbacks here
-          // since open-sse is a portable workspace without direct fs/os access.
-          "X-Msh-Device-Name": (providerSpecificData?.deviceName as string) || osTypeStr,
-          "X-Msh-Os-Version": (providerSpecificData?.osVersion as string) || osTypeStr,
-        },
-        body: params,
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // Detect unrecoverable errors
-      try {
-        const parsed = JSON.parse(errorText);
-        const errorCode = parsed?.error;
-        if (errorCode === "invalid_grant" || errorCode === "invalid_request") {
-          log?.error?.(
-            "TOKEN_REFRESH",
-            "Kimi Coding refresh token invalid. Re-authentication required.",
-            {
-              errorCode,
-            }
-          );
-          return { error: "unrecoverable_refresh_error", code: errorCode };
-        }
-      } catch {
-        // not JSON — fall through
-      }
-
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Kimi Coding token", {
-        status: response.status,
-        error: errorText.slice(0, 200),
-      });
-      return null;
-    }
-
-    const tokens = await response.json();
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Kimi Coding token", {
-      hasNewAccessToken: !!tokens.access_token,
-      hasNewRefreshToken: !!tokens.refresh_token,
-      expiresIn: tokens.expires_in,
-    });
-
-    return {
-      accessToken: tokens.access_token,
-      refreshToken: tokens.refresh_token || refreshToken,
-      expiresIn: tokens.expires_in,
-      tokenType: tokens.token_type,
-      scope: tokens.scope,
-    };
-  } catch (error) {
-    log?.error?.(
-      "TOKEN_REFRESH",
-      `Network error refreshing Kimi Coding token: ${error instanceof Error ? error.message : String(error)}`
-    );
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for GitLab Duo OAuth tokens.
- * Token URL is instance-specific; resolves from providerSpecificData.baseUrl.
- * Uses PKCE authorization_code flow initially but refresh_token grant does NOT
- * require code_verifier — only client_id + refresh_token.
- * On invalid_grant (revoked/expired refresh token) returns the unrecoverable sentinel.
- */
-export async function refreshGitLabDuoToken(
-  refreshToken: string,
-  providerSpecificData: Record<string, unknown> | null | undefined,
-  log: RefreshLogger,
-  proxyConfig: unknown = null
-) {
-  if (!refreshToken) {
-    log?.warn?.("TOKEN_REFRESH", "No refresh token for GitLab Duo");
-    return null;
-  }
-
-  const baseUrl = resolveGitLabOAuthBaseUrl(providerSpecificData);
-  const endpoints = buildGitLabOAuthEndpoints(baseUrl);
-  const tokenUrl = endpoints.tokenUrl;
-
-  // client_id from providerSpecificData (stored at login) or fall back to PROVIDERS config
-  const clientId =
-    (providerSpecificData?.clientId as string) ||
-    PROVIDERS["gitlab-duo"]?.clientId ||
-    process.env.GITLAB_DUO_OAUTH_CLIENT_ID ||
-    process.env.GITLAB_OAUTH_CLIENT_ID ||
-    "";
-
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(tokenUrl, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-          Accept: "application/json",
-        },
-        body: buildFormParams({
-          grant_type: "refresh_token",
-          refresh_token: refreshToken,
-          client_id: clientId,
-        }),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // Detect unrecoverable token — GitLab returns standard OAuth2 error codes.
-      try {
-        const errorBody = JSON.parse(errorText);
-        const errorCode = errorBody.error;
-        if (errorCode === "invalid_grant" || errorCode === "invalid_request") {
-          log?.error?.(
-            "TOKEN_REFRESH",
-            "GitLab Duo refresh token invalid. Re-authentication required.",
-            {
-              errorCode,
-            }
-          );
-          return { error: "unrecoverable_refresh_error", code: errorCode };
-        }
-      } catch {
-        // not JSON — fall through
-      }
-
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh GitLab Duo token", {
-        status: response.status,
-        error: errorText.slice(0, 200),
-      });
-      return null;
-    }
-
-    const tokens = await response.json();
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed GitLab Duo token", {
-      hasNewAccessToken: !!tokens.access_token,
-      hasNewRefreshToken: !!tokens.refresh_token,
-      expiresIn: tokens.expires_in,
-    });
-
-    return {
-      accessToken: tokens.access_token,
-      refreshToken: tokens.refresh_token || refreshToken,
-      expiresIn: tokens.expires_in,
-    };
-  } catch (error) {
-    log?.error?.(
-      "TOKEN_REFRESH",
-      `Network error refreshing GitLab Duo token: ${error instanceof Error ? error.message : String(error)}`
-    );
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for Claude OAuth tokens
- */
-export async function refreshClaudeOAuthToken(refreshToken, log, proxyConfig: unknown = null) {
-  try {
-    // Standard OAuth2 token refresh uses form-urlencoded (not JSON)
-    const params = buildFormParams({
-      grant_type: "refresh_token",
-      refresh_token: refreshToken,
-      client_id: PROVIDERS.claude.clientId,
-    });
-
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(OAUTH_ENDPOINTS.anthropic.token, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-          Accept: "application/json",
-          "anthropic-beta": "oauth-2025-04-20",
-        },
-        body: params.toString(),
-      })
-    );
-
-    if (!response.ok) {
-      // Read + classify the body ONCE, shape-agnostic. A proxy/MITM can deliver
-      // the invalid_grant 400 as a JSON string, a double-encoded string, a
-      // nested {error:{code}}, or raw text — all must yield the sentinel so the
-      // HealthCheck deactivates instead of looping every 60s.
-      const { rawText, code } = await readRefreshErrorBody(response);
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Claude OAuth token", {
-        status: response.status,
-        error: rawText.slice(0, 300),
-      });
-      if (code === "invalid_grant" || code === "invalid_request") {
-        return { error: "unrecoverable_refresh_error", code };
-      }
-      return null;
-    }
-
-    const tokens = await response.json();
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Claude OAuth token", {
-      hasNewAccessToken: !!tokens.access_token,
-      hasNewRefreshToken: !!tokens.refresh_token,
-      expiresIn: tokens.expires_in,
-    });
-
-    return {
-      accessToken: tokens.access_token,
-      refreshToken: tokens.refresh_token || refreshToken,
-      expiresIn: tokens.expires_in,
-    };
-  } catch (error) {
-    log?.error?.("TOKEN_REFRESH", `Network error refreshing Claude token: ${error.message}`);
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for Google providers (Gemini, Antigravity)
- */
-export async function refreshGoogleToken(
-  refreshToken,
-  clientId,
-  clientSecret,
-  log,
-  proxyConfig: unknown = null
-) {
-  const response = await runWithProxyContext(proxyConfig, () =>
-    fetch(OAUTH_ENDPOINTS.google.token, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        Accept: "application/json",
-      },
-      body: buildFormParams({
-        grant_type: "refresh_token",
-        refresh_token: refreshToken,
-        client_id: clientId,
-        client_secret: clientSecret,
-      }),
-    })
-  );
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    log?.error?.("TOKEN_REFRESH", "Failed to refresh Google token", {
-      status: response.status,
-      error: errorText.slice(0, 200),
-    });
-
-    // Detect unrecoverable token (invalid_grant = revoked / expired refresh token)
-    try {
-      const errorBody = JSON.parse(errorText);
-      if (errorBody.error === "invalid_grant") {
-        log?.error?.("TOKEN_REFRESH", "Google refresh token invalid. Re-authentication required.", {
-          provider: "google",
-        });
-        return { error: "unrecoverable_refresh_error", code: "invalid_grant" };
-      }
-    } catch {
-      // not JSON — fall through
-    }
-
-    return null;
-  }
-
-  const tokens = await response.json();
-
-  log?.info?.("TOKEN_REFRESH", "Successfully refreshed Google token", {
-    hasNewAccessToken: !!tokens.access_token,
-    hasNewRefreshToken: !!tokens.refresh_token,
-    expiresIn: tokens.expires_in,
-  });
-
-  return {
-    accessToken: tokens.access_token,
-    refreshToken: tokens.refresh_token || refreshToken,
-    expiresIn: tokens.expires_in,
-  };
-}
-
-export async function refreshQwenToken(refreshToken, log, proxyConfig: unknown = null) {
-  const endpoint = OAUTH_ENDPOINTS.qwen.token;
-
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(endpoint, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-          Accept: "application/json",
-        },
-        body: buildFormParams({
-          grant_type: "refresh_token",
-          refresh_token: refreshToken,
-          client_id: PROVIDERS.qwen.clientId,
-        }),
-      })
-    );
-
-    if (response.status === 200) {
-      const tokens = await response.json();
-
-      log?.info?.("TOKEN_REFRESH", "Successfully refreshed Qwen token", {
-        hasNewAccessToken: !!tokens.access_token,
-        hasNewRefreshToken: !!tokens.refresh_token,
-        expiresIn: tokens.expires_in,
-      });
-
-      return {
-        accessToken: tokens.access_token,
-        refreshToken: tokens.refresh_token || refreshToken,
-        expiresIn: tokens.expires_in,
-        providerSpecificData: tokens.resource_url
-          ? { resourceUrl: tokens.resource_url }
-          : undefined,
-      };
-    } else {
-      const errorText = await response.text().catch(() => "");
-
-      // Detect unrecoverable invalid_request (expired/revoked refresh token or bad client_id)
-      let errorCode = null;
-      try {
-        const parsed = JSON.parse(errorText);
-        errorCode = parsed?.error;
-      } catch {
-        // not JSON, ignore
-      }
-
-      if (errorCode === "invalid_request" || errorCode === "invalid_grant") {
-        log?.error?.(
-          "TOKEN_REFRESH",
-          "Qwen refresh token is invalid or expired. Re-authentication required.",
-          {
-            status: response.status,
-            errorCode,
-          }
-        );
-        return { error: "unrecoverable_refresh_error", code: errorCode };
-      }
-
-      log?.warn?.("TOKEN_REFRESH", `Error with Qwen endpoint`, {
-        status: response.status,
-        error: errorText,
-      });
-    }
-  } catch (error) {
-    log?.warn?.("TOKEN_REFRESH", `Network error trying Qwen endpoint`, {
-      error: error.message,
-    });
-  }
-
-  log?.error?.("TOKEN_REFRESH", "Failed to refresh Qwen token");
-  return null;
-}
-
-/**
- * Specialized refresh for Codex (OpenAI) OAuth tokens.
- * OpenAI uses rotating (one-time-use) refresh tokens.
- * Returns { error: 'unrecoverable_refresh_error', code } when the token has already been
- * consumed or is invalid, so callers can stop retrying and request re-authentication.
- */
-export async function refreshCodexToken(refreshToken, log, proxyConfig: unknown = null) {
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(OAUTH_ENDPOINTS.openai.token, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-          Accept: "application/json",
-        },
-        // Body intentionally omits `scope`. RFC 6749 §6 makes scope optional on a
-        // refresh_token grant (the server reuses the originally-granted scope when
-        // absent). Including `scope` causes Auth0 (which OpenAI Codex OAuth is
-        // built on) to treat the request as a re-scope, which can invalidate
-        // sibling refresh_token families on the same client_id. Matches the
-        // pattern used by ndycode/codex-multi-auth, the only known tool that
-        // sustains multiple Codex accounts without cross-invalidation.
-        body: buildFormParams({
-          grant_type: "refresh_token",
-          refresh_token: refreshToken,
-          client_id: PROVIDERS.codex.clientId,
-        }),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // Detect unrecoverable "refresh_token_reused" or "invalid_grant" error from OpenAI
-      // This means the token was already consumed or has expired.
-      // Retrying with the same token will never succeed.
-      let errorCode = null;
-      try {
-        const parsed = JSON.parse(errorText);
-        errorCode =
-          parsed?.error?.code || (typeof parsed?.error === "string" ? parsed.error : null);
-      } catch {
-        // not JSON, ignore
-      }
-
-      if (
-        errorCode === "refresh_token_reused" ||
-        errorCode === "invalid_grant" ||
-        errorCode === "token_expired" ||
-        errorCode === "invalid_token"
-      ) {
-        log?.error?.(
-          "TOKEN_REFRESH",
-          "Codex refresh token already used or invalid. Re-authentication required.",
-          {
-            status: response.status,
-            errorCode,
-          }
-        );
-        return { error: "unrecoverable_refresh_error", code: errorCode };
-      }
-
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Codex token", {
-        status: response.status,
-        error: errorText,
-      });
-      return null;
-    }
-
-    const tokens = await response.json();
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Codex token", {
-      hasNewAccessToken: !!tokens.access_token,
-      hasNewRefreshToken: !!tokens.refresh_token,
-      expiresIn: tokens.expires_in,
-    });
-
-    return {
-      accessToken: tokens.access_token,
-      refreshToken: tokens.refresh_token || refreshToken,
-      expiresIn: tokens.expires_in,
-    };
-  } catch (error) {
-    log?.error?.("TOKEN_REFRESH", `Network error refreshing Codex token: ${error.message}`);
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for Kiro (AWS CodeWhisperer) tokens
- * Supports both AWS SSO OIDC (Builder ID/IDC) and Social Auth (Google/GitHub)
- */
-export async function refreshKiroToken(
-  refreshToken,
-  providerSpecificData,
-  log,
-  proxyConfig: unknown = null
-) {
-  try {
-    const authMethod = providerSpecificData?.authMethod;
-    const clientId = providerSpecificData?.clientId;
-    const clientSecret = providerSpecificData?.clientSecret;
-    const region = providerSpecificData?.region;
-
-    // AWS SSO OIDC (Builder ID or IDC)
-    // If clientId and clientSecret exist, assume AWS SSO OIDC (default to builder-id if authMethod not specified).
-    // Exception: imported social tokens (authMethod === "imported") carry a freshly-registered
-    // clientId/clientSecret but their refresh token is Kiro-social-issued — the isolated OIDC client
-    // cannot refresh it, so they must fall through to the social auth path (#2467).
-    if (clientId && clientSecret && authMethod !== "imported") {
-      const endpoint = `https://oidc.${region || "us-east-1"}.amazonaws.com/token`;
-
-      const response = await runWithProxyContext(proxyConfig, () =>
-        fetch(endpoint, {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Accept: "application/json",
-          },
-          body: JSON.stringify({
-            clientId: clientId,
-            clientSecret: clientSecret,
-            refreshToken: refreshToken,
-            grantType: "refresh_token",
-          }),
-        })
-      );
-
-      if (!response.ok) {
-        const errorText = await response.text();
-
-        // AWS SSO OIDC uses {"__type": "InvalidGrantException"} error format (not standard OAuth2).
-        let awsErrorType: string | undefined;
-        try {
-          const awsError = JSON.parse(errorText);
-          awsErrorType = awsError.__type || awsError.error;
-        } catch {
-          // not JSON
-        }
-
-        // If the refresh token itself is expired/revoked, no amount of re-registration helps.
-        if (
-          awsErrorType === "InvalidGrantException" ||
-          awsErrorType === "ExpiredTokenException" ||
-          awsErrorType === "invalid_grant"
-        ) {
-          log?.error?.(
-            "TOKEN_REFRESH",
-            "Kiro AWS refresh token expired/invalid. Re-authentication required.",
-            { awsErrorType }
-          );
-          return { error: "unrecoverable_refresh_error", code: awsErrorType };
-        }
-
-        // Client credentials may be expired/invalid (DB import, TTL expiry, browser conflict).
-        // Re-register a fresh OIDC client and retry once before giving up (#2524).
-        log?.warn?.(
-          "TOKEN_REFRESH",
-          "Kiro OIDC refresh failed, attempting client re-registration...",
-          { status: response.status, error: errorText.slice(0, 200) }
-        );
-
-        try {
-          const resolvedRegion = region || "us-east-1";
-          const regEndpoint = `https://oidc.${resolvedRegion}.amazonaws.com/client/register`;
-          const regRes = await runWithProxyContext(proxyConfig, () =>
-            fetch(regEndpoint, {
-              method: "POST",
-              headers: { "Content-Type": "application/json", Accept: "application/json" },
-              body: JSON.stringify({
-                clientName: "kiro-oauth-client",
-                clientType: "public",
-                scopes: [
-                  "codewhisperer:completions",
-                  "codewhisperer:analysis",
-                  "codewhisperer:conversations",
-                ],
-                grantTypes: ["urn:ietf:params:oauth:grant-type:device_code", "refresh_token"],
-                issuerUrl: "https://identitycenter.amazonaws.com/ssoins-722374e8c3c8e6c6",
-              }),
-            })
-          );
-
-          if (regRes.ok) {
-            const newClient = await regRes.json();
-            const retryRes = await runWithProxyContext(proxyConfig, () =>
-              fetch(endpoint, {
-                method: "POST",
-                headers: { "Content-Type": "application/json", Accept: "application/json" },
-                body: JSON.stringify({
-                  clientId: newClient.clientId,
-                  clientSecret: newClient.clientSecret,
-                  refreshToken: refreshToken,
-                  grantType: "refresh_token",
-                }),
-              })
-            );
-
-            if (retryRes.ok) {
-              const retryTokens = await retryRes.json();
-              log?.info?.("TOKEN_REFRESH", "Kiro refresh recovered via client re-registration", {
-                hasNewAccessToken: !!retryTokens.accessToken,
-                expiresIn: retryTokens.expiresIn,
-              });
-              return {
-                accessToken: retryTokens.accessToken,
-                refreshToken: retryTokens.refreshToken || refreshToken,
-                expiresIn: retryTokens.expiresIn,
-                _newClientId: newClient.clientId,
-                _newClientSecret: newClient.clientSecret,
-                _newClientSecretExpiresAt: newClient.clientSecretExpiresAt,
-              };
-            }
-          }
-        } catch (reRegErr) {
-          log?.warn?.("TOKEN_REFRESH", "Kiro client re-registration fallback failed", {
-            error: String(reRegErr),
-          });
-        }
-
-        log?.error?.("TOKEN_REFRESH", "Failed to refresh Kiro AWS token", {
-          status: response.status,
-          error: errorText.slice(0, 200),
-        });
-        return null;
-      }
-
-      const tokens = await response.json();
-
-      log?.info?.("TOKEN_REFRESH", "Successfully refreshed Kiro AWS token", {
-        hasNewAccessToken: !!tokens.accessToken,
-        expiresIn: tokens.expiresIn,
-      });
-
-      return {
-        accessToken: tokens.accessToken,
-        refreshToken: tokens.refreshToken || refreshToken,
-        expiresIn: tokens.expiresIn,
-      };
-    }
-
-    // Social Auth (Google/GitHub) - use Kiro's refresh endpoint
-    const tokenUrl = PROVIDERS.kiro.tokenUrl;
-    if (!tokenUrl) {
-      log?.error?.("TOKEN_REFRESH", "Missing Kiro token endpoint");
-      return null;
-    }
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch(tokenUrl, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Accept: "application/json",
-        },
-        body: JSON.stringify({
-          refreshToken: refreshToken,
-        }),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-
-      // Also check for AWS-style errors on the social auth path (Kiro may relay them)
-      try {
-        const awsError = JSON.parse(errorText);
-        const awsErrorType = awsError.__type || awsError.error;
-        if (
-          awsErrorType === "InvalidGrantException" ||
-          awsErrorType === "ExpiredTokenException" ||
-          awsErrorType === "invalid_grant"
-        ) {
-          log?.error?.(
-            "TOKEN_REFRESH",
-            "Kiro social refresh token expired/invalid. Re-authentication required.",
-            {
-              awsErrorType,
-            }
-          );
-          return { error: "unrecoverable_refresh_error", code: awsErrorType };
-        }
-      } catch {
-        // not JSON — fall through
-      }
-
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Kiro social token", {
-        status: response.status,
-        error: errorText.slice(0, 200),
-      });
-      return null;
-    }
-
-    const tokens = await response.json();
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Kiro social token", {
-      hasNewAccessToken: !!tokens.accessToken,
-      expiresIn: tokens.expiresIn,
-    });
-
-    return {
-      accessToken: tokens.accessToken,
-      refreshToken: tokens.refreshToken || refreshToken,
-      expiresIn: tokens.expiresIn,
-    };
-  } catch (error) {
-    log?.error?.("TOKEN_REFRESH", `Network error refreshing Kiro token: ${error.message}`);
-    return null;
-  }
-}
-
-/**
- * Specialized refresh for Qoder OAuth tokens
- */
-export async function refreshQoderToken(refreshToken, log, proxyConfig: unknown = null) {
-  if (!OAUTH_ENDPOINTS.qoder.token || !PROVIDERS.qoder.clientId || !PROVIDERS.qoder.clientSecret) {
-    log?.warn?.(
-      "TOKEN_REFRESH",
-      "Qoder OAuth refresh skipped: browser OAuth is not configured in this environment"
-    );
-    return null;
-  }
-
-  const basicAuth = btoa(`${PROVIDERS.qoder.clientId}:${PROVIDERS.qoder.clientSecret}`);
-
-  const response = await runWithProxyContext(proxyConfig, () =>
-    fetch(OAUTH_ENDPOINTS.qoder.token, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        Accept: "application/json",
-        Authorization: `Basic ${basicAuth}`,
-      },
-      body: buildFormParams({
-        grant_type: "refresh_token",
-        refresh_token: refreshToken,
-        client_id: PROVIDERS.qoder.clientId,
-        client_secret: PROVIDERS.qoder.clientSecret,
-      }),
-    })
-  );
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    log?.error?.("TOKEN_REFRESH", "Failed to refresh Qoder token", {
-      status: response.status,
-      error: errorText,
-    });
-    const code = extractOAuthErrorCode(errorText);
-    if (code === "invalid_grant" || code === "invalid_request") {
-      return { error: "unrecoverable_refresh_error", code };
-    }
-    return null;
-  }
-
-  const tokens = await response.json();
-
-  log?.info?.("TOKEN_REFRESH", "Successfully refreshed Qoder token", {
-    hasNewAccessToken: !!tokens.access_token,
-    hasNewRefreshToken: !!tokens.refresh_token,
-    expiresIn: tokens.expires_in,
-  });
-
-  return {
-    accessToken: tokens.access_token,
-    refreshToken: tokens.refresh_token || refreshToken,
-    expiresIn: tokens.expires_in,
-  };
-}
-
-/**
- * Specialized refresh for GitHub Copilot OAuth tokens
- */
-export async function refreshGitHubToken(refreshToken, log, proxyConfig: unknown = null) {
-  const response = await runWithProxyContext(proxyConfig, () =>
-    fetch(OAUTH_ENDPOINTS.github.token, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        Accept: "application/json",
-      },
-      body: buildFormParams({
-        grant_type: "refresh_token",
-        refresh_token: refreshToken,
-        client_id: PROVIDERS.github.clientId,
-        client_secret: PROVIDERS.github.clientSecret,
-      }),
-    })
-  );
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    log?.error?.("TOKEN_REFRESH", "Failed to refresh GitHub token", {
-      status: response.status,
-      error: errorText,
-    });
-    const code = extractOAuthErrorCode(errorText);
-    if (code === "invalid_grant" || code === "invalid_request") {
-      return { error: "unrecoverable_refresh_error", code };
-    }
-    return null;
-  }
-
-  const tokens = await response.json();
-
-  log?.info?.("TOKEN_REFRESH", "Successfully refreshed GitHub token", {
-    hasNewAccessToken: !!tokens.access_token,
-    hasNewRefreshToken: !!tokens.refresh_token,
-    expiresIn: tokens.expires_in,
-  });
-
-  return {
-    accessToken: tokens.access_token,
-    refreshToken: tokens.refresh_token || refreshToken,
-    expiresIn: tokens.expires_in,
-  };
-}
-
-/**
- * Refresh GitHub Copilot token using GitHub access token
- */
-export async function refreshCopilotToken(githubAccessToken, log, proxyConfig: unknown = null) {
-  try {
-    const response = await runWithProxyContext(proxyConfig, () =>
-      fetch("https://api.github.com/copilot_internal/v2/token", {
-        headers: getGitHubCopilotRefreshHeaders(`token ${githubAccessToken}`),
-      })
-    );
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      log?.error?.("TOKEN_REFRESH", "Failed to refresh Copilot token", {
-        status: response.status,
-        error: errorText,
-      });
-      return null;
-    }
-
-    const data = await response.json();
-
-    log?.info?.("TOKEN_REFRESH", "Successfully refreshed Copilot token", {
-      hasToken: !!data.token,
-      expiresAt: data.expires_at,
-    });
-
-    return {
-      token: data.token,
-      expiresAt: data.expires_at,
-    };
-  } catch (error) {
-    log?.error?.("TOKEN_REFRESH", "Error refreshing Copilot token", {
-      error: error.message,
-    });
-    return null;
-  }
-}
-
-/**
- * Get access token for a specific provider (internal, does the actual work)
- */
-async function _getAccessTokenInternal(provider, credentials, log, proxyConfig: unknown = null) {
-  switch (provider) {
-    case "gemini":
-    case "gemini-cli":
-    case "antigravity":
-    case "agy":
-      return await refreshGoogleToken(
-        credentials.refreshToken,
-        PROVIDERS[provider].clientId,
-        PROVIDERS[provider].clientSecret,
-        log,
-        proxyConfig
-      );
-
-    case "claude":
-      return await refreshClaudeOAuthToken(credentials.refreshToken, log, proxyConfig);
-
-    case "codex":
-      return await refreshCodexToken(credentials.refreshToken, log, proxyConfig);
-
-    case "qwen":
-      return await refreshQwenToken(credentials.refreshToken, log, proxyConfig);
-
-    case "qoder":
-      return await refreshQoderToken(credentials.refreshToken, log, proxyConfig);
-
-    case "github":
-      return await refreshGitHubToken(credentials.refreshToken, log, proxyConfig);
-
-    case "kiro":
-    case "amazon-q":
-      return await refreshKiroToken(
-        credentials.refreshToken,
-        credentials.providerSpecificData,
-        log,
-        proxyConfig
-      );
-
-    case "cline":
-      return await refreshClineToken(credentials.refreshToken, log, proxyConfig);
-
-    case "kimi-coding":
-      return await refreshKimiCodingToken(
-        credentials.refreshToken,
-        credentials.providerSpecificData,
-        log,
-        proxyConfig
-      );
-
-    case "gitlab-duo":
-      return await refreshGitLabDuoToken(
-        credentials.refreshToken,
-        credentials.providerSpecificData,
-        log,
-        proxyConfig
-      );
-
-    case "windsurf":
-    case "devin-cli":
-      return await refreshWindsurfToken(
-        credentials.refreshToken,
-        credentials.providerSpecificData,
-        log,
-        proxyConfig
-      );
-
-    default:
-      // Fallback to generic OAuth refresh for unknown providers
-      return refreshAccessToken(provider, credentials.refreshToken, credentials, log, proxyConfig);
-  }
-}
-
-/**
- * Whether a provider has a supported refresh path in this service.
  */
 export function supportsTokenRefresh(provider) {
   const explicitlySupported = new Set([
